refactor(modules): move EPUB modules to src/modules/epub/
Folder structure now mirrors the navigator split:
  src/modules/
    ReaderModule.ts, ModuleHost.ts, ModuleRegistry.ts,
    ModuleAccessors.ts, NavigatorFeatureMap.ts, interfaces.ts
    highlight/          (shared highlighter, used by EPUB modules)
    epub/               (EPUB-specific modules)
    pdf/                (PDF-specific modules)

Single-file modules flattened into epub/ directly (no subfolder).
Multi-file modules kept their folders: epub/TTS/, epub/search/,
epub/mediaoverlays/.

All import paths updated across the codebase (~100+ lines).
No logic changes.
