+++ src/modules/epub/CitationModule.ts
